shellpilot/requirements.txt
pyte>=0.8
pywinpty>=2.0; sys_platform == 'win32'
pygame>=2.0
pyyaml>=6.0
openai>=1.0
python-dotenv>=1.0
imageio>=2.0
imageio-ffmpeg>=0.4
numpy>=1.20
pillow>=9.0
